+++ src/components/InputArea/InputArea.js
@@ -12,21 +12,43 @@ const Wrapper = styled.div`
 `;
 
 class InputArea extends React.Component {
-
+  
+  renderInitial() {
+    return (
+      <ChildComponents.Initial
+        handleNewClick={this.props.handleNewClick}
+      />
+    );
+  }
+  
+  renderNew() {
+    return (
+      <ChildComponents.New
+        charactersCountLabel={this.props.charactersCountLabel}
+        text={this.props.text}
+        handleSaveClick={this.props.handleSaveClick}
+        handleCancelClick={this.props.handleCancelClick}
+        handleTextareaChange={this.props.handleTextareaChange}
+      />
+    );
+  }
+  
+  renderEdit() {
+    return (
+      <ChildComponents.Edit />
+    );
+  }
+  
+  renderSaving() {
+    return <ChildComponents.Saving />;
+  }
+  
   renderChildComponent() {
     switch(this.props.status) {
-      case 'initial': return <ChildComponents.Initial />;
-      case 'new':
-        return (
-          <ChildComponents.New
-            charactersCountLabel={this.props.charactersCountLabel}
-          />
-        );
-      case 'edit':
-        return (
-          <ChildComponents.Edit />
-        );
-      case 'saving': return <ChildComponents.Saving />;
+      case 'initial': return this.renderInitial();
+      case 'new': return this.renderNew();
+      case 'edit': return this.renderEdit();
+      case 'saving': return this.renderSaving();
       default: return null;
     }
   }
@@ -48,6 +70,11 @@ InputArea.propTypes = {
     'saving'
   ]).isRequired,
   charactersCountLabel: PropTypes.string.isRequired,
+  text: PropTypes.string.isRequired,
+  handleNewClick: PropTypes.func.isRequired,
+  handleCancelClick: PropTypes.func.isRequired,
+  handleSaveClick: PropTypes.func.isRequired,
+  handleTextareaChange: PropTypes.func.isRequired,
 };
 
 export default InputArea;