+++ src/components/index.js
@@ -6,5 +6,6 @@ export { default as Board } from './Board/Board';
 export { default as Button } from './Button/Button';
 export { default as InputArea } from './InputArea/InputArea';
 export { default as Message } from './Message/Message';
+export { default as Spinner } from './Spinner/Spinner';
 export { default as TextArea } from './TextArea/TextArea';
 export { default as Thread } from './Thread/Thread';