+++ src/constants/index.js
@@ -1,1 +1,7 @@
 export const MAX_MESSAGE_LENGTH = 240;
+
+/** NOTE
+// API_BASE_URL should depend on the environment
+// Use dotenv to define a REACT_APP_API_BASE_URL environment variable
+**/
+export const API_BASE_URL = 'http://localhost:3004';