+++ src/redux/__tests__/inputArea.test.js
@@ -1,4 +1,6 @@
+import { Map } from 'immutable';
 import reducer, { INITIAL_STATE, actions, selectors } from '../inputArea';
+import { types as messagesTypes } from '../messages';
 import { MAX_MESSAGE_LENGTH } from '../../constants'; 
 
 describe('inputArea duck', () => {
@@ -22,5 +24,26 @@ describe('inputArea duck', () => {
       expect(selectors.getText(nextState)).toEqual('');
     });
 
+  });
+  
+  it('sets status to "saving" on a SAVE', () => {
+    const state = Map({ status: 'new' });
+    const expectedState = Map({ status: 'saving' });
+    const action = { type: messagesTypes.SAVE };
+    expect(reducer(state, action)).toEqual(expectedState);
+  });
+  
+  it('sets status to "initial" on SAVE_SUCCEEDED', () => {
+    const state = Map({ status: 'saving' });
+    const action = { type: messagesTypes.SAVE_SUCCEEDED };
+    const nextState = reducer(state, action); 
+    expect(selectors.getStatus(nextState)).toEqual('initial');
+  });
+  
+  it('sets text to an empty string on SAVE_SUCCEEDED', () => {
+    const state = Map({ text: 'whatever' });
+    const action = { type: messagesTypes.SAVE_SUCCEEDED };
+    const nextState = reducer(state, action); 
+    expect(selectors.getText(nextState)).toEqual('');
   });
 });