+++ src/redux/index.js
@@ -1,3 +1,4 @@
 // Reducers. Keep in alphabetical order.
+export { default as app } from './app';
 export { default as inputArea } from './inputArea';
 export { default as messages } from './messages';