+++ src/redux/messages.js
@@ -8,6 +8,12 @@ export const types = {
   ADD: 'messages/ADD',
   REMOVE: 'messages/REMOVE',
   UPDATE: 'messages/UPDATE',
+  FETCH: 'messages/FETCH',
+  FETCH_SUCCEEDED: 'messages/FETCH_SUCCEEDED',
+  FETCH_FAILED: 'messages/FETCH_FAILED',
+  SAVE: 'messages/SAVE',
+  SAVE_SUCCEEDED: 'messages/SAVE_SUCCEEDED',
+  SAVE_FAILED: 'messages/SAVE_FAILED',
 };
 
 // Reducer
@@ -27,11 +33,18 @@ const reduceUpdate = (state, action) => {
   return state;
 };
 
+const reduceFetchSucceeded = (state, action) => {
+  return List(action.payload.data);
+};
+
 export default function reducer(state = INITIAL_STATE, action = {}) {
   switch(action.type) {
     case types.ADD: return state.push(new Message(action.payload.message));
     case types.REMOVE: return reduceRemove(state, action);
     case types.UPDATE: return reduceUpdate(state, action);
+    case types.FETCH_SUCCEEDED: return reduceFetchSucceeded(state, action);
+    case types.SAVE_SUCCEEDED:
+      return state.push(new Message(action.payload.data));
     default: return state;
   }
 }
@@ -53,10 +66,49 @@ const update = (id, text) => ({
   payload: { id, text }
 });
 
+const fetch = () => ({
+  type: types.FETCH
+});
+
+const fetchSucceeded = data => ({
+  type: types.FETCH_SUCCEEDED,
+  payload: { data }
+});
+
+const fetchFailed = error => ({
+  type: types.FETCH_FAILED,
+  payload: { error }
+});
+
+const save = ({text, parentId, author}) => ({
+  type: types.SAVE,
+  payload: {
+    text,
+    parentId,
+    author
+  }
+});
+
+const saveSucceeded = data => ({
+  type: types.SAVE_SUCCEEDED,
+  payload: { data }
+});
+
+const saveFailed = error => ({
+  type: types.SAVE_FAILED,
+  payload: { error }
+});
+
 export const actions = {
   add,
   remove,
   update,
+  fetch,
+  fetchSucceeded,
+  fetchFailed,
+  save,
+  saveSucceeded,
+  saveFailed,
 };
 
 // Selectors